Add password specifier to setup

+++ manager/trackmon_db_setup.sh
@@ -11,13 +11,18 @@ else
   exit 1
 fi
 
-whiptail --passwordbox "Enter your database password. REMEMBER THIS!" 8 78 --title "Trackmon Database Setup" 2> password
-PASSWORD=$(cat password)
-rm password
+
 echo "Creating new user trackmon"
 sudo adduser --disabled-login --gecos 'Trackmon' trackmon
 echo "Trying to create postgres user trackmon"
-sudo -u postgres psql -d template1 -c "CREATE USER trackmon CREATEDB PASSWORD '$PASSWORD';"
+if (whiptail --title "Trackmon Database Setup" --yesno "Do you want to set a database password? Required when trackmon accesses the database remotely!" 8 78) then
+  whiptail --passwordbox "Enter your database password. REMEMBER THIS!" 8 78 --title "Trackmon Database Setup" 2> password
+  PASSWORD=$(cat password)
+  rm password
+  sudo -u postgres psql -d template1 -c "CREATE USER trackmon CREATEDB PASSWORD '$PASSWORD';"
+else
+  sudo -u postgres psql -d template1 -c "CREATE USER trackmon CREATEDB;"
+fi
 echo "Trying to create database trackmon_server_production"
 sudo -u postgres psql -d template1 -c "CREATE DATABASE trackmon_server_production OWNER trackmon;"
